Made editor form its own component
